Removed all web dependencies from homotopy-graphics, moved style.rs back into graphics, made homotopy-model depend on homotopy-graphics.
