+++ homotopy-graphics/Cargo.toml
@@ -12,11 +12,9 @@ build = true
 [dependencies]
 homotopy-common = { path = "../homotopy-common" }
 homotopy-core = { path = "../homotopy-core" }
-homotopy-model = { path = "../homotopy-model" }
 # TODO(@doctorn) remove if/when we switch to GL-backed 2D rendering
 euclid = "0.22.7"
 itertools = "0.10.3"
-js-sys = "0.3.59"
 log = "0.4.16"
 lyon_algorithms = "1.0.1"
 lyon_geom = "1.0.1"
@@ -28,23 +26,6 @@ serde = { features = ["derive"], version = "1.0.144" }
 serde_json = "1.0.85"
 thiserror = "1.0.32"
 ultraviolet = "0.9.0"
-wasm-bindgen = "0.2.82"
-yew = "0.19.3"
-
-[dependencies.web-sys]
-version = "0.3.58"
-features = [
-  "HtmlCanvasElement",
-  "WebGlBuffer",
-  "WebGlFramebuffer",
-  "WebGlProgram",
-  "WebGlRenderbuffer",
-  "WebGl2RenderingContext",
-  "WebGlShader",
-  "WebGlTexture",
-  "WebGlUniformLocation",
-  "WebGlVertexArrayObject",
-]
 
 [dev-dependencies]
 cargo-husky = { features = ["precommit-hook", "run-cargo-check", "run-cargo-test", "run-cargo-fmt", "run-cargo-clippy"], version = "1.5.0" }
