+++ homotopy-graphics/src/lib.rs
@@ -46,7 +46,7 @@ pub mod manim;
 pub mod parity;
 pub mod path_util;
 pub mod stl;
-pub use homotopy_model::graphics::style;
+pub mod style;
 pub mod subdivision;
 pub mod svg;
 pub mod tikz;
