+++ homotopy-graphics/src/style.rs
