--- homotopy-model/src/graphics.rs
@@ -1,1 +1,0 @@
-pub mod style;
