+++ homotopy-model/src/lib.rs
@@ -1,2 +1,1 @@
-pub mod graphics;
 pub mod model;
