+++ homotopy-model/src/model/serialize.rs
@@ -7,6 +7,7 @@ use homotopy_core::{
     serialize::{Key, Store},
     Diagram,
 };
+use homotopy_graphics::style::Color;
 use im::Vector;
 use obake::AnyVersion;
 
@@ -14,7 +15,6 @@ use super::{
     proof::{generators::GeneratorInfo, FolderInfo, SignatureItem, View},
     Signature, Workspace,
 };
-use crate::graphics::style::Color;
 
 #[obake::versioned]
 #[obake(version("0.1.0"))]
